Update worker configuration.
Now that server and worker modes are implied by the presence of a :server and :worker key, we can't use values associated with :server because this will require the server namespace and state.

Along the way, it became apparent that the worker is consuming all available messages, so a worker channel is defined that limits the number of consumer unack'd messages to one.