Use clj-http for downloading non-local files.
The advantage of clj-http is that it will retry failed requests automatically (much like the Cassandra driver will for failed queries). We may still want to use a general-purpose library for retrying failures, but not when something like this already exists.

The util ns now provides a multi-method for retrieving a local file and a remote file (via http).