Update tile-spec geometry calculation.
The shift_x and shift_y value only took into account a pixel shift because source data was previously guaranteed to have upper-left/lower-right points that aligned to a grid that was an even multiple of tile sizes.

Now that sources conform to a grid that is not an even multiple of the tile-size (in the projection coordinate system) the value stored in shift_x and shift_y must be changed.