Revisit tile-spec creation. (#16)
Added a helper to dev/user.clj for producing a tile-spec.json that can be used with POST /landsat/tile-specs
Side-step a problem creating data_mask from JSON; this will have to be revisited later. The current schema expects a data_mask map of int -> text, but JSON data is parsed as str -> str.